phase 4 finished
finished issues #24, #29, #31, #32

Met with accessibility consultant. Made requested modifications except for <form> and <fieldset> on questionnaire.html

We agreed <fieldset> accessibility issues (no legend) could be ignored because the information/semantic structure is still clear.

I couldn't modify <form> to <div role=form> because it would've added extra work on javascript side (froze the entire dynamics when I tried), so probably that's an issue for the next version of the website. For now, I added an accessible name to it.